TDD: Fix mis-quoted named section.
The regular expression would only match named sections that had matched
quotes, making the "trim only the left quote" else block impossible to
hit, leading to an entry in the dictionary with whatever section was
last (an empty string or the previous section).  I think this new
behaviour is closer to the intended interpretation, even though it is
now impossible to have a section name containing a closing square
bracket, but I think that's a good enough tradeoff.