Fix for incorrect paths for installation checks
Incorrect paths are added when the installations are added through the shellPaths. 
Effectively, the name "Git.exe" is also added along as the directory path. 

Also, to note is that recent versions of Git portable versions have 2 "git.exe" files, one in the `bin` directory, and the other in the `cmd` directory.