Finishing touches on GitHub.Authentication UX
Inverted the dependency on GitHub.Authentication vs Micorosft.Alm.Authentication, such that GitHub now depends on Microsoft.Alm, this is necissary as the Microsoft.Alm is a pure library while the GitHub code is an application in its own right.

Made Microsoft.Alm.CredentialHelper depend on GitHub.Authentication, which only makes sense since it needs the support of the new binary to perform its necissary duties when interacting with GitHub. Additionally, it makes the copy/deploy story more simplified as only a single csproj is copying binaries to the .\Deploy folder.