Add NTLM detection to `BasicAuthentication`.
Add a new `WwwAuthenticateHelper` to detect authentication type support of hosts via HTTP headers, specifically the 'WWW-Authenticate' header.

Add dynamic callback support to the `BasicAuthentication` class, such that modal or basic prompts can be chosen when the class is created and invoked when `AcquireCredentials` is called. This, correctly, moves the logic for deciding what to call to the configuration stage and the the execution logic to the execution state.

Add NTLM detection to `BasicAuthentication.AcquireCredentials` method, which wholly independent from the `GetCredentials` method now.