Refactor VSTS specific code to separate library.
Refactor the project so that all code specific to Visual Studio Team Services lives in a new "Microsoft.Vsts.Authentication" library. This leaves all of the "common" code behind in "Microsoft.Alm.Authentication" library.

The long-term goal here is to make VSTS, GitHub, and Atlassian code share as much surface area as possible, thus reducing repetition and bugs.

Required a number of changes in build, deployment, and packaging steps; specifically with regards to the new "Microsoft.Vsts.Authentication.dll" binary which is now necessary to authenticate with TFS and VSTS hosted repositories.