Move doc build script to root of sln.
Having documentation generation in the 'Cli-CredentialHelper' directory doesn't make a lot of sense. Instead, it should centralized to the solution as it already generates docs for both Credential Manager and Ask Pass.

Added more clairity to the error message provided when 'build-docs.cmd' fails due the dependency: Pandoc is either not installed or cannot be found on $path.