Implement explicit support for authentication of repositories hosted at "bitbucket.org". It provides a Bitbucket branded UI with support for Basic Authentication credentials and those users who have activated 2FA on their account to continue to use HTTPS Git URLs.
2FA support is provided via Bitbucket's OAuth implementation, users are able to run the OAuth authorization process and the GCM can retrieve and store OAuth tokens and return them to Git in a format acceptable to Bitbucket.

Bitbucket authentication support is implemented as a peer of the existing Microsoft VSTS and GitHub implementations, in the Bitbucket.Authentication project. The implementation, as far as possible, mirrors that of the GitHub support. In an attempt to lay the ground work of support for other hosts in the future some shared GUI concepts have been pulled out into the GUI-Shared shared code project. However only the new Bitbucket implementation makes use of them, the GitHub implementation has NOT been refactored to use these shared components at this time.

Further information can be found in the BITBUCKET_README.md file within the repository.