Make askpass work when username is known.
Make it so that Cli-Askpass functions correctly when the username is known. This includes keeping the username portion of the `TargetUri.ActualUri`.

Make Cli-Askpass try harder to parse credentials from queries passed to it from Git.

Make Cli-Askpass correctly captured basic credentials, for readback at a later time.

Make basic credential modal prompt only ask for password when the username is already known.