Authenticate gist repos properly
Gists are backed by git repositories hosted on GitHub.com, but the host name for these repos are "gist.github.com" and not "github.com".

This confuses the GCM as it tries to retrieve credentials for gist.github.com in this case. This change normalizes such URLs so that any operation against "gist.github.com" goes against "github.com".

This is special cased against gist.github.com rather than global because you never know if GitHub might use subdomains for other purposes in the future.

Fixes #495