installer: add Installer project.
Completes the initial effort to enable binary + installer signing support.

Using the MicroBuild solution, signing is now enabled for any environment which is setup to support it; and a pass-through operation for environments where it is not.

Final outputs will be in the Deploy/ folder, which will only contain the Microsoft.Alm.Authentication.{version}.nupkg, gcmw-v{version}.zip, and GCMW-{version}.exe files after a successful build.