ms.alm.git: adopt disposing pattern
Because `IDisposable` implementations' `Dispose` method can be called from either runtime code or a type finalizer. When called from a finalizer, the state of references is unknown/unreliable. Therefore, `Dispose` needs to differentiate between finalizers and other callers; and ought to avoid updating any references it doesn't need to when called from a finalizer.