Ms.Vsts.Auth: Improve credential deletion logic.
When Git asks for a VSTS PAT to be erased, it is not always because the credentials are actually invalid. Often it is due to network issues (flakiness, WI-FI disconnected, etc.). Therefore is better to validate that the credentials have gone invalid before deleting them.

Validation before deletion is a good idea for the reason that authentication is often a laborious task; often requiring multi-factor steps such as answering a phone or getting a code from email. A flaky network should not cause a user to waste so much time re-authenticating.

When credentials are invalidated, there's no good way to know if it is because they have expired or another reason. A recently common reason is accounts migrating between Azure tenants. When this happens the tenant cache become a nuance because the user cannot authenticate correctly because the cause points them at the wrong authority.

Therefore, when credentials are purged the related authority ought to be as well.