cli-shared: Enumerate config and environ keys.
Instead of keeping each configuration key and each environment key as a separately declared global constant, define a set of possible keys as an enumeration. Then define a look up of type -> key for configuration keys, and for environment keys.

Update the `LoadOperationArguments` and related functions to use the new schema. This is code clean up and 'correctness' change.