cli-shared: Remove the concept to of actual vs query URI values.
At some point in the past, there was a valid reason to keep query and actual URI separate concepts within `TargetUri`. However, that time has passed and having both concepts is confusion. This patch remedies the confusion by removing the `ActualUri` concept, leaving `TargerUri` with `QueryUri` and `ProxyUri`.

Modify `OperationArguments` to track credentials as a `Credential` type field. Changed the logic so that `OperationArguments.Username` is set via initialization and no longer by `.SetCredentials`. This enables the GCM to differentiate between username passed in as an argument from Git (assume part of target URL) and arguments passed in by user as part of answering a prompt.