bitbucket: Adopt service-provider model & remove 'Microsoft.Alm.Git' from solution.
The current GCM design is incredibly difficult to test, and much of it is impossible to unit-test. The source of the difficulty
resides in the number of hard-links, static, and global resources and API the project binds to.

Introduction of the service-provider model will begin to enable test authors to substitute external and static API with mocks,
etc. Types now can extend `Microsoft.Alm.BaseType` which exposes service provded by an instance of the `RuntimeContext` type.
With `RuntimeContext` type as the provider, all access to file-system, network, and where services are instance base; avoiding
global static instances (which are impossible to test correctly).

Any dependency on 'Microsoft.Alm.Authentication' required a dependency on 'Micrsoft.Alm.Git'. The dual dependency requirement
is due to excessive exposure of 'Microsoft.Alm.Git' types in 'Micrsoft.Alm.Authentication' public API. This fact is a strong
indication that two assemblies ought to be a single assembly.

This change moves all of the *.Git types into the *.Authentication assembly.