installer: Remove 'Microsoft.Alm.Git' from solution.
Any dependency on 'Microsoft.Alm.Authentication' required a dependency on 'Micrsoft.Alm.Git'. The dual dependency requirement
is due to excessive exposure of 'Microsoft.Alm.Git' types in 'Micrsoft.Alm.Authentication' public API. This fact is a strong
indication that two assemblies ought to be a single assembly.

This change moves all of the *.Git types into the *.Authentication assembly.