Add README instructions for installing in MSYS2
I think MSYS2 is a pretty common base for a unixy dev environment on windows, and GCM works well with the MSYS2-packaged `git`, but it wasn't obvious to me at first what the expected way to install it was. Hopefully this README chunk will clarify it for future folks.

I also moved the "manual installation" subsection into the "Download and Install" section rather than the "How To Use" section.