authentication: Remove appliation knowledge.
The 'Microsoft.Alm.Authentication' library ought to not know about the GCM directly, and its knowledge of the "GCM_TRACE" environment variable is a violation of this assertion.

Remove the knowledge and use of "GCM_TRACE" from the library. In the next commit, the logic will be moved to the application layer.