authentication: Rename "file system" to "storage" #1.
Since a file system is a very specific kind of storage, and there's no reason any of 'Microsoft.Alm.Authentication' API
should care about the kind of storage, using "file system" could be a misnomer. Instead using "storage" as the type name makes
more sense.

This change is the first in a series which renames all types, interfaces, fields, paramaters, and properties from "file system"
to "storage", starting with the `FileSystem` type itself.