cli: change parent hwnd to envvar.
Instead of passing the parent window handle (aka hwnd) via the command line, the value ought to be passed via environment variable. This is because the primary use-case is for Git to invoke the GCM.

In the case of IDE invoking GIt, using environment variables provides a way for the IDE to specify which window ought to be the GCM's parent, despite Git being an intermediary.