installer: purge old files before copying new ones.
As the list of files to be deployed changes over time, the installer needs to keep a list of files to remove (in addition to the list of files to copy).

This way, old, out-of-date files can be removed during installations. This is important to avoid any version conflicts that might occur due to side-by-side or unexpected assembly load issues.