build: consolidate signing in installer project.
Consolodate all signing operations in the Installer project. Signing take a long time, stretching builds which take a minute or two to complete into a build which takes tens of minutes to complete. The majority of the added time is due to signing request submission overhead. Therefore, batching requests for signing minimizes the additional time.

Correctly consolidating signing means adding a new Components sub-project to the Installer project. Given that signing submissions can only happen once per project, and that the individual binaries need to be signed prior to 'setup.exe' and 'gcmw-*.zip' files can be generated; a prior signing submission must happen - thus the Components project. Additionally, there is no longer a reason to retain the "MicroBuild" dependencies in the other projects, no a reason to include the "MicroBuild" targets and props.