alm: avoid cycles by keep account
When building a hierarchical list after enumerating process, it is possible to encounter a cycle in the process hierarchy reported by Windows. This will cause the previous solution (linked list) to allocate the repeating sequence until the system runs out of memory.

This change adds an accouting step before each entry is added to the list, if the entry has already been seen the loop is terminated and the entry omitted.