vsts: validate credentials, handle anonymous accounts.
VSTS has made a series of changes to their Identity service.

One such change is that the '_apis/connectiondata' always returns either "200 OK" or "404 Not Found". This means the returned content must be examined to determine if the credentials are valida or not.

Another change is that requests lacking "accepts", "accepts-encoding", and "cache-control" headers are aways given the 'anonymous user' response.

 - [x] Provide a method of detection for anonymous accounts during credential validation.
 - [x] Ensure that credential validation requests can handle HTTP cookies as part of their response flow.
 - [x] Add "accepts", "accepts-encoding", and "cache-control" headers to all network requests.
 - [x] Provide handling of "content-encoding" content of "gzip" and "deflate".