docs: update readme to indicate archiving of project
Update the readme for the project to indicate it is no longer being
maintained, and that GCM Core is the replacement.

Also mention the future removal of password-based access to GitHub's API
for creating tokens (that GCM for Windows uses) - this will result in
GCM for Windows being dead-in-the-water in the near future.